removed duplicate code, cleanup, some documentation
