+++ src/main/java/com/doomhamsters/CardType.java
@@ -1,3 +1,4 @@
+package com.doomhamsters;
 
 /** Represents a single card in the draw deck. */
 public enum CardType {
