+++ src/main/java/com/doomhamsters/DrawError.java
@@ -1,3 +1,5 @@
+package com.doomhamsters;
+
 /** Error response sent to the requesting client on a failed draw attempt. */
 
 public final class DrawError {
@@ -6,7 +8,7 @@ public final class DrawError {
   private final String playerId;
   private final String reason;
 
-  /** Constructor for the DrawError. */
+  /** Constructor for the default DrawError. */
   public DrawError(String gameId, String playerId, String reason) {
     this.gameId = gameId;
     this.playerId = playerId;
