+++ src/test/java/com/doomhamsters/DrawErrorTest.java
@@ -1,3 +1,4 @@
+package com.doomhamsters;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.Test;
