+++ src/test/java/com/doomhamsters/DrawRequestTest.java
@@ -1,5 +1,5 @@
 
-
+package com.doomhamsters;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.Test;
