+++ src/test/java/com/doomhamsters/GameStateTest.java
@@ -1,4 +1,4 @@
-
+package com.doomhamsters;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
